+++ apps/meteor/client/providers/ServerProvider.tsx
@@ -1,5 +1,4 @@
 import type { Serialized } from '@rocket.chat/core-typings';
-import { Emitter } from '@rocket.chat/emitter';
 import type { Method, PathFor, OperationParams, OperationResult, UrlParams, PathPattern } from '@rocket.chat/rest-typings';
 import type {
 	ServerMethodName,
@@ -59,58 +58,16 @@ const callEndpoint = <TMethod extends Method, TPathPattern extends PathPattern>(
 
 const uploadToEndpoint = (endpoint: PathFor<'POST'>, formData: any): Promise<UploadResult> => sdk.rest.post(endpoint as any, formData);
 
-const getStream = <N extends StreamNames, K extends StreamKeys<N>>(
-	streamName: N,
-	_options?: {
-		retransmit?: boolean | undefined;
-		retransmitToSelf?: boolean | undefined;
-	},
-): ((eventName: K, callback: (...args: StreamerCallbackArgs<N, K>) => void) => () => void) => {
-	return (eventName, callback): (() => void) => {
-		return sdk.stream(streamName, [eventName], callback as (...args: any[]) => void).stop;
-	};
-};
-
-const ee = new Emitter<Record<string, void>>();
-
-const events = new Map<string, () => void>();
-
-const getSingleStream = <N extends StreamNames, K extends StreamKeys<N>>(
-	streamName: N,
-	_options?: {
-		retransmit?: boolean | undefined;
-		retransmitToSelf?: boolean | undefined;
-	},
-): ((eventName: K, callback: (...args: StreamerCallbackArgs<N, K>) => void) => () => void) => {
-	const stream = getStream(streamName);
-	return (eventName, callback): (() => void) => {
-		ee.on(`${streamName}/${eventName}`, callback);
-
-		const handler = (...args: any[]): void => {
-			ee.emit(`${streamName}/${eventName}`, ...args);
-		};
-
-		const stop = (): void => {
-			// If someone is still listening, don't unsubscribe
-			ee.off(`${streamName}/${eventName}`, callback);
-
-			if (ee.has(`${streamName}/${eventName}`)) {
-				return;
-			}
-
-			const unsubscribe = events.get(`${streamName}/${eventName}`);
-			if (unsubscribe) {
-				unsubscribe();
-				events.delete(`${streamName}/${eventName}`);
-			}
-		};
-
-		if (!events.has(`${streamName}/${eventName}`)) {
-			events.set(`${streamName}/${eventName}`, stream(eventName, handler));
-		}
-		return stop;
-	};
-};
+const getStream =
+	<N extends StreamNames>(
+		streamName: N,
+		_options?: {
+			retransmit?: boolean | undefined;
+			retransmitToSelf?: boolean | undefined;
+		},
+	) =>
+	<K extends StreamKeys<N>>(eventName: K, callback: (...args: StreamerCallbackArgs<N, K>) => void): (() => void) =>
+		sdk.stream(streamName, [eventName], callback).stop;
 
 const contextValue = {
 	info,
@@ -119,7 +76,6 @@ const contextValue = {
 	callEndpoint,
 	uploadToEndpoint,
 	getStream,
-	getSingleStream,
 };
 
 const ServerProvider: FC = ({ children }) => <ServerContext.Provider children={children} value={contextValue} />;