+++ apps/meteor/client/views/room/contextualBar/Threads/hooks/useThreadMainMessageQuery.ts
@@ -18,10 +18,18 @@ type NotifyRoomRidDeleteMessageBulkEvent = {
 	ignoreDiscussion: boolean;
 	ts: FieldExpression<Date>;
 	users: string[];
+	ids?: string[]; // message ids have priority over ts
+	showDeletedStatus?: boolean;
 };
 
 const createDeleteCriteria = (params: NotifyRoomRidDeleteMessageBulkEvent): ((message: IMessage) => boolean) => {
-	const query: Query<IMessage> = { ts: params.ts };
+	const query: Query<IMessage> = {};
+
+	if (params.ids) {
+		query._id = { $in: params.ids };
+	} else {
+		query.ts = params.ts;
+	}
 
 	if (params.excludePinned) {
 		query.pinned = { $ne: true };