+++ apps/meteor/ee/app/authorization/server/callback.ts
@@ -12,10 +12,36 @@ License.onInstall(() => {
 		callbacks.priority.HIGH,
 		'validateUserRoles',
 	);
-	callbacks.add('afterSaveUser', () => License.shouldPreventAction('activeUsers'), callbacks.priority.HIGH, 'validateUserRoles');
-	callbacks.add('afterDeleteUser', () => License.shouldPreventAction('activeUsers'), callbacks.priority.HIGH, 'validateUserRoles');
+	callbacks.add(
+		'afterSaveUser',
+		async (user) => {
+			await License.shouldPreventAction('activeUsers');
 
-	callbacks.add('afterDeactivateUser', () => License.shouldPreventAction('activeUsers'), callbacks.priority.HIGH, 'validateUserStatus');
+			return user;
+		},
+		callbacks.priority.HIGH,
+		'validateUserRoles',
+	);
+	callbacks.add(
+		'afterDeleteUser',
+		async (user) => {
+			await License.shouldPreventAction('activeUsers');
+
+			return user;
+		},
+		callbacks.priority.HIGH,
+		'validateUserRoles',
+	);
+
+	callbacks.add(
+		'afterDeactivateUser',
+		async (user) => {
+			await License.shouldPreventAction('activeUsers');
+			return user;
+		},
+		callbacks.priority.HIGH,
+		'validateUserStatus',
+	);
 
 	callbacks.add(
 		'beforeActivateUser',