+++ packages/instance-status/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocket.chat/instance-status
 
+## 0.0.26
+
+### Patch Changes
+
+- @rocket.chat/models@0.0.26
+
+## 0.0.25
+
+### Patch Changes
+
+- @rocket.chat/models@0.0.25
+
 ## 0.0.24
 
 ### Patch Changes